requirements.txt
groq
langchain
langchain-groq
langchain_community
chromadb
python-dotenv
streamlit
pydantic
pypdf
langchain-core
typing-extensions>=4.5.0
numpy>=1.22.0
pandas>=1.3.0
tqdm>=4.65.0
regex>=2023.6.3
sentence-transformers>=2.2.0
torch>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0

